// src/typings/api.ts
export interface ApiAddOnProductLink {
    productId: number
    productName: string
}

export interface ApiAddOnProductSimple extends ApiAddOnProductLink {
    parameters: ApiAddOnProductParameter[]
}

export interface ApiAddOnProduct extends ApiAddOnProductSimple {
    dependencies: AddOnProductDependency
    childAddOnProducts: ApiAddOnProduct[]
}

export interface AddOnProductDependency {
    excludes: ApiAddOnProductLink[]
    requires: ApiAddOnProductLink[]
}

export interface ApiAddOnProductParameter {
    parameterId: number
    name: string
    required: boolean
}

export interface ApiSubscriptionProductAddOnProducts {
    families: ApiSubscriptionProductAddOnProductsFamily[]
}

export interface ApiSubscriptionProductAddOnProductsFamily {
    name: string
    familyGroupType: 'EXTRA_SIM_CARDS' | 'UNKNOWN'
    addOnProducts: ApiAddOnProduct[]
}